requirements.txt
fastapi
uvicorn
streamlit
python-dotenv
langchain
langchain-openai
langchain-community
langchain-experimental
requests
